Fix Fixnum deprecation warning in ruby 2.4
Sumamry
-------

Ruby 2.4 unified Fixnum and Bignum into a single class Integer and
added deprecation warnings for explicit uses of Fixnum and Bignum. As
such using this gem with newer versions of ruby pollutes the logs with
deprecation warnings. This attempts to remove the deprecation warning
by referencing the correct numeric class